+++ src/main/java/net/oneandone/lavender/modules/ModuleProperties.java
@@ -144,11 +144,14 @@ public class ModuleProperties extends PropertiesBase {
                 String scmSvn = "scm:svn:";
 
                 scmurlProd = (String) properties.remove(prefix);
-                scmurlDevel = eatOpt(properties, prefix + ".devel", scmurlProd);
-                if (scmurlProd.startsWith(scmSvn) && !scmurlDevel.startsWith(scmSvn)) { // TODO
-                    scmurlDevel = scmSvn + scmurlDevel;
-                    LOG.warn("fixed devel url: " + scmurlDevel + ". Please update this module to the latest parent pom to fix this warning.");
+                if (!scmurlProd.startsWith(scmSvn)) { // work-around missing scmsvn prefix when released with prerelease plugin
+                    scmurlProd = scmSvn + scmurlProd;
                 }
+                scmurlDevel = eatOpt(properties, prefix + ".devel", scmurlProd);
+                if (!scmurlDevel.startsWith(scmSvn)) { // work-around missing scmsvn prefix when released with prerelease plugin
+                    scmurlDevel = scmSvn + scmurlDevel;
+                }
+
                 tag = eatOpt(properties, prefix + ".revision", "-1");
                 scmsrc = eatLegacySvnSource(properties, prefix, source);
                 scmsrc = fallback(scmurlProd, scmsrc);